Revert setting global registery remover to null and how shared memory is destroyed.
This change sets the listener back to a NOP instead of a NULL, which fixes problems around waking up after a suspend.

This also reverts the changes of handling shared memory.
Shared memory is no longer created only if the shared memory doesn't exists, but can be attached to by another main instance at any time.
This change,however, unmaps the memory of the instance instead of ulinking the shared memory object.